# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bzcom CRM — Customer Request Management System API
  description: >
    REST API quản lý yêu cầu khách hàng (bug / feature / inquiry) cho Bzcom.
    Mọi response dùng envelope thống nhất `{ status, message, data }`.
    Xác thực bằng JWT Bearer token (trừ login, refresh, logout và đăng ký CLIENT).
  version: 1.0.0
  contact:
    name: KITS Hanoi OJT 2026 — Bzcom Developer Team
servers:
  - url: http://localhost:8080
    description: Local (Docker Compose)
...
paths:
...
  /api/requests:
    post:
      tags: [Request]
      summary: Tạo yêu cầu (CLIENT only). HIGH → tự sinh alert cho ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestCreateRequest' }
      responses:
        '201':
          description: Tạo request thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RequestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [Request]
      summary: Danh sách request (phân trang / sort / filter tổ hợp; phạm vi theo role)
      description: >
        ADMIN thấy tất cả; CLIENT chỉ thấy request mình tạo;
        DEVELOPER chỉ thấy request được gán cho mình.
      parameters:
        - { name: page,     in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size,     in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
        - name: sort
          in: query
          description: 'field,direction; field hợp lệ: id, title, category, priority, status, clientId, assignedDeveloperId, createdAt, updatedAt'
          schema: { type: string, example: 'createdAt,desc' }
        - { name: status,   in: query, schema: { $ref: '#/components/schemas/RequestStatus' } }
        - { name: category, in: query, schema: { $ref: '#/components/schemas/Category' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/Priority' } }
        - { name: keyword,  in: query, schema: { type: string }, description: 'tìm trong title/description' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PageRequestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/requests/{id}/assign:
    patch:
      tags: [Request]
      summary: Gán developer (ADMIN only). auto=true → thuật toán tự chọn.
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignRequest' }
      responses:
        '200':
          description: Gán thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RequestResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: Không có DEVELOPER khả dụng để auto-assign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/requests/{id}/status:
    patch:
      tags: [Request]
      summary: Cập nhật trạng thái; request phải được gán developer và transition phải hợp lệ → 409
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusUpdateRequest' }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RequestResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Request chưa được gán developer, transition không hợp lệ hoặc optimistic lock conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
              example: { status: 409, message: 'Invalid status transition: DONE -> IN_PROGRESS', data: null }
  /api/requests/{id}/history:
    get:
      tags: [Request]
      summary: Lịch sử thay đổi trạng thái của request
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Danh sách history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/HistoryResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Dữ liệu không hợp lệ (validation)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 400, message: 'title: must not be blank', data: null }
    Unauthorized:
      description: Chưa đăng nhập hoặc token không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 401, message: 'Unauthorized', data: null }
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 403, message: 'Forbidden', data: null }
    NotFound:
      description: Không tìm thấy resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 404, message: 'Request not found: 999', data: null }
    Conflict:
      description: Xung đột (email trùng / transition sai / optimistic lock)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 409, message: 'Email already exists', data: null }
  schemas:
    ApiResponse:
      type: object
      required: [status, message, data]
      properties:
        status:  { type: integer, example: 200 }
        message: { type: string, example: success }
        data:    { nullable: true }
    RequestStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, DONE]
    Category:
      type: string
      enum: [BUG, FEATURE, INQUIRY]
    Priority:
      type: string
      enum: [HIGH, MEDIUM, LOW]
...
    RequestCreateRequest:
      type: object
      additionalProperties: false
      required: [title, category, priority]
      properties:
        title:       { type: string, minLength: 1, maxLength: 200, example: 'Login fails' }
        description: { type: string, maxLength: 4000, example: 'When I click login, page shows 500 error' }
        category:    { $ref: '#/components/schemas/Category' }
        priority:    { $ref: '#/components/schemas/Priority' }
...
    RequestResponse:
      type: object
      properties:
        id:          { type: integer, format: int64 }
        title:       { type: string }
        description: { type: string, nullable: true }
        category:    { $ref: '#/components/schemas/Category' }
        priority:    { $ref: '#/components/schemas/Priority' }
        status:      { $ref: '#/components/schemas/RequestStatus' }
        clientId:    { type: integer, format: int64 }
        assignedDeveloperId: { type: integer, format: int64, nullable: true }
        version:     { type: integer, minimum: 0, description: 'Gửi lại ở mọi PATCH để phát hiện stale update' }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    PageRequestResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/RequestResponse' }
        page:          { type: integer, example: 0 }
        size:          { type: integer, example: 10 }
        totalElements: { type: integer, example: 42 }
        totalPages:    { type: integer, example: 5 }
    AssignRequest:
      oneOf:
        - type: object
          additionalProperties: false
          required: [auto, expectedVersion]
          properties:
            auto: { type: boolean, enum: [true] }
            expectedVersion: { type: integer, minimum: 0, example: 0 }
        - type: object
          additionalProperties: false
          required: [auto, developerId, expectedVersion]
          properties:
            auto: { type: boolean, enum: [false] }
            developerId: { type: integer, format: int64, minimum: 1 }
            expectedVersion: { type: integer, minimum: 0, example: 0 }
    StatusUpdateRequest:
      type: object
      required: [status, expectedVersion]
      properties:
        status: { $ref: '#/components/schemas/RequestStatus' }
        memo:   { type: string, maxLength: 255, example: 'start working' }
        expectedVersion: { type: integer, minimum: 0, example: 0 }
    HistoryResponse:
      type: object
      properties:
        id:         { type: integer, format: int64 }
        requestId:  { type: integer, format: int64 }
        changedBy:  { type: integer, format: int64 }
        fromStatus:
          allOf: [{ $ref: '#/components/schemas/RequestStatus' }]
          nullable: true
        toStatus:
          allOf: [{ $ref: '#/components/schemas/RequestStatus' }]
          nullable: true
        memo:       { type: string, nullable: true }
        changedAt:  { type: string, format: date-time }
...
security:
  - bearerAuth: []